python-crawler/requirements.txt
scrapy==2.11.0
psycopg2-binary
